+++ src/main/java/com/kch/likelion/LikeLionKing/user/JdbcUserRepository.java
@@ -1,4 +1,4 @@
-package com.kch.likelion.LikeLionKing.user.repository;
+package com.kch.likelion.LikeLionKing.user;
 
 import com.kch.likelion.LikeLionKing.user.domain.User;
 import com.kch.likelion.LikeLionKing.user.domain.UserRole;
